requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
clickhouse-connect==0.7.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
numpy>=1.24.0
pandas>=2.0.0
pyproj>=3.6.0
scipy>=1.11.0
